refactor(capi): extract CommandScreen + PointerInput to CAPI_screen.h (todo 2.3)
Moved verbatim from CAPI.cpp (todo said CAPI_screen.cpp; both classes are
fully inline so the extraction is header-only — a .cpp would be an empty
anchor). utf8_codepoint_count moved with them (static -> inline).
CAPI.cpp: 2316 -> 2058 lines. CAPI_internal.h fwd-decls updated.
Verified: full rebuild, 45/45 ctest, symbol freeze 110.
